Move -c flag to end of r10k command
Apparently, in the latest version of r10k the flag needs to be passed AFTER the subcommand. If it comes first (as it did prior to this commit) it is silently ignored.

That might be a bug in r10k but calling the flag last gets it to work for now.